compat: <unistd.h>: make X_OK a no-op on Windows
There's nothing to check via access() on Windows for directory
permissions, and I have no clue if the underlying _access() ignores
unknown mode flags, so let's play it safe. The worst case is that
access() will be too optimistic, which wouldn't really hart because of
TOCTOU races, anyway. In other words, if anything breaks by this patch,
then it's broken already.