Allow building on platforms without `struct tm.tm_gmtoff`
I'm not happy about this because the results are non-intuitive, but that
looks like something which originates in libyang's attempts to use local
TZ everywhere instead of normalizing to UTC time.

But hey, this only matters on a platform which cannot "easily" represent
timezones in `struct tm`, so as long as we mark the returned value with
"I cannot really vouch for the TZ", this might even be standard-compliant.