MSVC: there's no mmap() on Windows
There's a compatibility library, the mman-win32, which implement a
mmap-like semantics. However, on Windows one cannot map a read-only file
into memory for a length that is larger than the size of that file,
and you also cannot overwrite an existing mapping.

The easiest course of action is to just read the file into memory and be
done with it.

See-also: https://stackoverflow.com/a/46014637/2245623