Use semicolon as a directory delimiter on Windows
Windows path names have ':' inside as a drive letter separator. The
widespread convention on that platform is to use ';' as a list separator
in file paths, so let's follow suite.

It looks like this multi-value searchdir is only used in `ly_ctx_new()`,
and that all other functions operate on one directory at a time.