Windows: tests: do not invoke syscalls on an invalid FDs
On Windows, calling fstat() on a FD that was already closed results in
calling the "invalid parameter handler". Set up on so that we can catch
this properly. I'm doing this instead of just commenting out everything
because I would prefer to keep that coverage to make sure we won't leak
the FD in question.