Windows: tests: do not invoke syscalls on an invalid FDs
On Windows, calling fstat() on a FD that was already closed results in
calling the "invalid parameter handler".

Originally I created a proper set up for this to make sure that the code
won't leak a FD, but Michal prefers a saner approach of not testing for
Windows idiosyncrasies, which makes perfect sense to me.