MSVC: tests: fix paths to utests and test plugins
On Unix, libyang plugins are not subject to RUNTIME_OUTPUT_DIRECTORY,
instead requiring LIBRARY_OUTPUT_DIRECTORY. On Windows, the MSVC
generator is a multi-config one, which means that it will happily put
files into an extra subdirectory of RUNTIME_OUTPUT_DIRECTORY. This is
super-hard to fix, so instead just query CMake for whatever the
generator thinks is suitable, and use *that*.

This means that we have to distinguish between "the test binary
directory", "directory with utests binaries" and "directory with test
plugin binaries", so while we're at that, make sure that these paths are
different than they used to be before this change -- simply to make sure
that all bugs will be visible even on single-config generators like make
or ninja. The tests still use the "binary dir" for things like
CMakeFiles (!), which is super implementation-specific, but at least now
the actual binaries live elsewhere.

Querying the xxx_OUTPUT_DIRECTORY, however, is not enough because that
property does *not* have the config-specific suffix. Generators to the
rescue, but that also means that the generated file suddenly has to be
generator-config specific as well, which means that we have to massage
the incldue path even more. Yay.

Also, on windows the test executables have the `.exe` suffix.