CI: start building on Windows
The differences compared to a POSIX build are too big even in the build
script, mainly because CMake considers the Visual Studio build file
generator a multiconfig one, and that means that all invocations of
cmake/ctest have to have a specification of the current config. There's
no way around that in CMake (and boy did I try!), so let's give up and
do everything from scratch.